refactor(adm/ADMTrees): split 1864-LOC monolith into adm/trees/ package
Split python/pdstools/adm/ADMTrees.py into a focused subpackage so the
6 type-ignore lines could be removed and so each concern can be read
on its own:

- adm/trees/_nodes.py    — Split, Node, parse_split, traversal helpers
- adm/trees/_model.py    — ADMTreesModel (kept whole; mixin-splitting
                           judged worse than the size win)
- adm/trees/_multi.py    — MultiTrees collection + datamart decoders
- adm/trees/_agb.py      — datamart-level AGB discovery namespace
- adm/trees/__init__.py  — public re-exports + deprecated ADMTrees(...)
                           polymorphic factory
- adm/ADMTrees.py        — backward-compat shim

Type-ignore delta: 6 -> 0 (requests stub now ships; per-branch cast for
_safe_condition_evaluate; getattr for pydot create_<format>).

Public API unchanged. All callers (datasets.py, ADMDatamart.py, every
test, both example notebooks) continue to use
`from pdstools.adm.ADMTrees import ...` via the shim. The
`pdstools.adm.ADMTrees` logger name is preserved (pinned in _model.py)
so the safe-eval logging test keeps binding correctly.

Tests: 81/81 in test_ADMTrees.py pass unchanged; full suite green.

---

Also folds in seven AGENTS.md additions distilled from the recent
multi-PR sweep, plus a new "gold-standard checklist" section that
codifies the bar this refactor (and prior ADMDatamart work) is held to:

- Polars filter(~bool_with_nulls) gotcha (#667).
- agg(...).any(ignore_nulls=False) for tri-state flags.
- Filter sentinel rows (e.g. AGB totals row) at the IO boundary.
- Cloud-IO classmethod testing pattern (boto3 + moto).
- Gold-standard checklist for top-level analyzer classes.
- Read the whole issue thread before designing the fix.
- Don't land plan-files-only PRs.
- Worktree rescue playbook for session restarts.
